refactor: use new plugin structure
